+++ dashboard-app/src/dashboard/ByActivity/index.tsx
@@ -13,9 +13,9 @@ import { CommunityBusinesses } from '../../api';
 import { DurationUnitEnum } from '../../types';
 import useRequest from '../../hooks/useRequest';
 import { useAggDataOnRes } from '../../hooks/useAggDataOnRes';
-import { tableType } from '../../util/dataManipulation/tableType';
-import { aggregatedToTableData } from '../../util/dataManipulation/aggregatedToTableData';
-import { downloadCsv } from '../../util/dataManipulation/downloadCsv';
+import { tableType } from '../dataManipulation/tableType';
+import { aggregatedToTableData } from '../dataManipulation/aggregatedToTableData';
+import { downloadCsv } from '../dataManipulation/downloadCsv';
 
 const DataTable = styled(_DataTable)`
   margin-top: 4rem;
