+++ dashboard-app/src/dashboard/dataManipulation/__tests__/aggregatedToTableData.test.ts
