+++ dashboard-app/src/dashboard/dataManipulation/__tests__/logsToAggregatedData.test.ts
