+++ dashboard-app/src/dashboard/dataManipulation/aggregatedToCsv.ts
